BE:fix: change data type of 'row' and 'col' in ClubAsset Entity

See merge request s08-final/S08P31A102!553

+++ backend/service/src/main/java/org/lightnsalt/hikingdom/service/info/dto/response/AssetInfoRes.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AssetInfoRes {
 	private String assetUrl;
-	private int row;
-	private int column;
+	private double row;
+	private double column;
 
 	public AssetInfoRes(ClubAsset clubAsset) {
 		this.assetUrl = clubAsset.getAsset().getAssetUrl();
